requirements.txt
# FastAPI and Web Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
python-dotenv==1.0.1

# OpenAI API (GPT-4o + DALL-E 3)
openai==1.58.1

# Image Processing and Computer Vision
pillow==11.0.0
opencv-python-headless==4.10.0.84

# PDF Generation
reportlab==4.2.5

# HTTP Requests and File Handling
requests==2.32.3
aiofiles==24.1.0

# AWS S3 Integration
boto3==1.36.14
botocore==1.36.14

# Data Validation and Serialization
pydantic==2.10.5
pydantic-settings==2.10.1
email-validator==2.2.0

# Database
sqlalchemy==2.0.42
alembic==1.16.4
psycopg2-binary==2.9.10

# Additional Utilities
typing-extensions==4.12.2
httpx==0.28.1

# Development and Testing (optional but useful)
pytest==8.3.4
pytest-asyncio==0.25.3

# Security and Performance
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OAuth and authentication
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
authlib==1.2.1

stripe==7.8.0
PyJWT==2.8.0
bcrypt==4.1.2
